requirements.txt
Django==2.2.4
django-crispy-forms==1.7.2
django-phone-field==1.8.0
pytz==2019.2
sqlparse==0.3.0
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.5
whitenoise==5.2.0